third_party/blink/web_tests/external/wpt/fetch/private-network-access/mixed-content-fetch.tentative.https.window_include=from-treat-as-public-expected.txt

This is a testharness.js-based test.
FAIL treat-as-public-address to private: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
FAIL treat-as-public-address to local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
Harness: the test ran to completion.
